Update Calendar and Event Pages
Clicking 'Events' in the sidebar now takes the user to the calendar.php
page to view the already created events.  From there, the user will be
able to navigate to the events.php page to create a new event, and back
to the calendar.php page to view the events in the database.